# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce de productos orgánicos",
        "description": "Documentación para la API de ecommerce de productos orgánicos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend-raices.vercel.app/api/api",
            "description": "Servidor de producción"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registra un usuario cliente",
                "operationId": "2d21e359fe383ed65e6eade8f537bb58",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "username",
                                    "firstname",
                                    "lastname",
                                    "password",
                                    "type_user",
                                    "phone_number",
                                    "gender",
                                    "birth_date"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Iniciar sesión",
                "operationId": "a18fa152592f54ccf7546ec09758a011",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "identifier",
                                    "password"
                                ],
                                "properties": {
                                    "identifier": {
                                        "description": "Nombre de usuario o email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "identifier": "olivia.perez@gmail.com",
                                "password": "12345678"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login satisfactorio"
                    },
                    "400": {
                        "description": "Los datos ingresados son incorrectos"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Cerrar sesión",
                "description": "Este endpoint cierra la sesión del usuario. Se requiere autenticación.",
                "operationId": "13d1b10ae1f04943de6d1707e9be6f62",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada"
                    },
                    "401": {
                        "description": "No autorizado, el usuario no está autenticado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}